@@ -2,18 +2,13 @@
   <m-scroll
     class="listview"
     :data="data"
-    ref="scroll"
+    ref="scrollRef"
     :probe-type="probeType"
     :listen-scroll="true"
     @scroll="handleScroll"
   >
-    <ul>
-      <li
-        class="list-group"
-        v-for="group of data"
-        :key="group.title"
-        ref="listGroup"
-      >
+    <ul ref="listGroupRef">
+      <li class="list-group" v-for="group of data" :key="group.title">
         <h2 class="list-group-title">{{ group.title }}</h2>
         <ul>
           <li
@@ -28,7 +23,7 @@
       </li>
     </ul>
     <div
-      class="list-letters"
+      class="list-lettersRef"
       @touchstart.stop.prevent="handleLetterTouchStart"
       @touchmove.stop.prevent="handleLetterTouchMove"
       @touchend.stop
@@ -36,19 +31,23 @@
       <ul>
         <li
           class="letter"
-          v-for="(item, index) of letters"
+          v-for="(item, index) of lettersRef"
           :key="item"
           :data-index="index"
-          :class="{ current: currentIndex === index }"
+          :class="{ current: currentIndexRef === index }"
         >
           {{ item }}
         </li>
       </ul>
     </div>
+    <div class="list-fixed" ref="fixed" v-show="fixedTitleRef">
+      <div class="fixed-title">{{ fixedTitleRef }}</div>
+    </div>
   </m-scroll>
 </template>
 
 <script>
+import { ref, computed, watch } from 'vue'
 import MScroll from '@components/m-scroll/index'
 import { getDomData } from '@/utils/dom'
 
@@ -67,109 +66,128 @@ export default {
       }
     }
   },
-  data() {
-    return {
-      currentIndex: 0,
-      scrollY: 0,
-      heightList: []
-    }
-  },
-  computed: {
-    letters() {
-      return this.data.map(v => {
+  setup(props, context) {
+    const currentIndexRef = ref(0)
+    const scrollYRef = ref(0)
+    const listGroupRef = ref(null)
+    let liListGroup = null
+    const scrollRef = ref(null)
+    let heightList = []
+    const touchObj = {}
+    const probeType = 3
+
+    const lettersRef = computed(() =>
+      props.data.map(v => {
         return v.title.substr(0, 1)
       })
-    }
-  },
-  watch: {
-    data() {
-      setTimeout(() => {
-        this._getHeightList()
-      }, 20)
-    },
-    scrollY(nv) {
+    )
+
+    const fixedTitleRef = computed(() => {
+      if (scrollYRef.value > 0) {
+        return ''
+      }
+      return props.data[currentIndexRef.value]
+        ? props.data[currentIndexRef.value].title
+        : ''
+    })
+
+    watch(
+      () => props.data,
+      () => {
+        setTimeout(() => {
+          _getHeightList()
+        }, 20)
+      }
+    )
+    watch(scrollYRef, nv => {
       const y = nv
       // 当滚动到顶部
       if (y > 0) {
-        this.currentIndex = 0
+        currentIndexRef.value = 0
         return
       }
       // 当滚动到中间部分
       // heightList 的长度比元素长度长 1 线段端点与线的关系
-      for (let i = 0, len = this.heightList.length - 1; i < len; i++) {
-        const left = this.heightList[i],
-          right = this.heightList[i + 1]
+      for (let i = 0, len = heightList.length - 1; i < len; i++) {
+        const left = heightList[i],
+          right = heightList[i + 1]
         if (-y >= left && -y < right) {
-          this.currentIndex = i
+          currentIndexRef.value = i
           return
         }
       }
       // 当滚动到最后 大于最后一个值的上限 -1 是下限
-      this.currentIndex = this.heightList.length - 2
-    }
-  },
-  created() {
-    this.touch = {}
-    this.probeType = 3
-  },
-  methods: {
-    handleLetterTouchStart(e) {
+      currentIndexRef.value = heightList.length - 2
+    })
+
+    function handleLetterTouchStart(e) {
       const touch = e.touches[0]
       const index = getDomData(e.target, 'index')
-      this.touch.y1 = touch.pageY
-      this.touch.startIndex = parseInt(index)
-      this._scrollTo(index)
-    },
-    handleLetterTouchMove(e) {
+      touchObj.y1 = touch.pageY
+      touchObj.startIndex = parseInt(index)
+      _scrollTo(index)
+    }
+    function handleLetterTouchMove(e) {
       const touch = e.touches[0]
-      this.touch.y2 = touch.pageY
-      const moveIndex = ((this.touch.y2 - this.touch.y1) / LETTER_HEIGHT) | 0
-      this._scrollTo(this.touch.startIndex + moveIndex)
-    },
+      touchObj.y2 = touch.pageY
+      const moveIndex = ((touchObj.y2 - touchObj.y1) / LETTER_HEIGHT) | 0
+      _scrollTo(touchObj.startIndex + moveIndex)
+    }
     // 使用函数方式，也可以使用 watch
     // handleScroll({ y }) {
     //   if (y > 0) {
-    //     this.currentIndex = 0
+    //     currentIndexRef.value = 0
     //     return
     //   }
-    //   for (let i = 0, len = this.heightList.length; i < len; i++) {
-    //     const left = this.heightList[i],
-    //       right = this.heightList[i + 1]
+    //   for (let i = 0, len = heightList.length; i < len; i++) {
+    //     const left = heightList[i],
+    //       right = heightList[i + 1]
 
     //     if (-y >= left && -y < right) {
-    //       this.currentIndex = i
+    //       currentIndexRef.value = i
     //       return
     //     }
     //   }
     //   // 大于最后一个值的上限 -1 是下限
-    //   this.currentIndex = this.heightList.length - 2
+    //   currentIndexRef.value = heightList.length - 2
     // },
-    handleScroll({ y }) {
-      this.scrollY = y
-    },
-    _getHeightList() {
-      console.log('_getHeightList')
-      this.heightList = []
+    function handleScroll({ y }) {
+      scrollYRef.value = y
+    }
+    function _getHeightList() {
+      heightList = []
       let height = 0
-      this.heightList.push(height)
-      for (const v of this.$refs.listGroup) {
+      heightList.push(height)
+      liListGroup = listGroupRef.value.children
+      for (const v of liListGroup) {
         height += v.clientHeight
-        this.heightList.push(height)
+        heightList.push(height)
       }
-    },
-    _scrollTo(index) {
+    }
+    function _scrollTo(index) {
       if (!index && index !== 0) {
         return
       }
       if (index < 0) {
         index = 0
-      } else if (index > this.heightList.length - 2) {
-        index = this.heightList.length - 2
+      } else if (index > heightList.length - 2) {
+        index = heightList.length - 2
       }
       // 当点击 letter 的时候 scroll 滚动不触发，则手动赋值 scrollY
       // 当前版本 bscroll 会触发
-      // this.scrollY = -this.heightList[index]
-      this.$refs.scroll.scrollToElem(this.$refs.listGroup[index])
+      // this.scrollY = -heightList[index]
+      scrollRef.value.scrollToElem(liListGroup[index])
+    }
+    return {
+      currentIndexRef,
+      lettersRef,
+      listGroupRef,
+      scrollRef,
+      fixedTitleRef,
+      probeType,
+      handleLetterTouchStart,
+      handleLetterTouchMove,
+      handleScroll
     }
   }
 }
@@ -193,6 +211,7 @@ export default {
       padding-left 20px
       font-size $font-size-small
       color $color-text-n
+      background-color $color-ef-ll
     .list-group-item
       display flex
       align-items center
@@ -205,7 +224,7 @@ export default {
         padding-left 20px
         color $color-text
         font-size $font-size-medium
-  .list-letters
+  .list-lettersRef
     z-index 30
     position absolute
     top 50%
@@ -224,4 +243,16 @@ export default {
       color $color-text-n
       &.current
         color $color-theme
+  .list-fixed
+    position absolute
+    top 0
+    left 0
+    width 100%
+    background-color $color-ef-ll
+    .fixed-title
+      height 30px
+      line-height 30px
+      padding-left: 20px
+      font-size $font-size-small
+      color $color-text-n
 </style>
