@@ -3,7 +3,7 @@ module.exports = {
   env: {
     node: true
   },
-  extends: ['plugin:vue/essential', '@vue/standard'],
+  extends: ['plugin:vue/vue3-essential', '@vue/standard'],
   parserOptions: {
     parser: 'babel-eslint'
   },
