@@ -19,7 +19,7 @@
           <ul>
             <li v-for="(item, index) of discList" :key="index" class="item">
               <div class="icon">
-                <img width="60" height="60" v-lazy="item.imgurl" />
+                <img width="60" height="60" :src="item.imgurl" />
               </div>
               <div class="text">
                 <h2 class="name" v-html="item.creator.name"></h2>
